# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Документация",
        "description": "API car booking",
        "version": "1"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Аутентификация"
                ],
                "summary": "Авторизация по email и паролю",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "manager@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешная авторизация",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Аутентификация"
                ],
                "summary": "Выход пользователя",
                "description": "Отзывает текущий токен доступа для выхода из системы",
                "operationId": "69281b12abb272c76871f19cb17ca563",
                "responses": {
                    "200": {
                        "description": "Успешный выход",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Успешный выход из системы"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Неавторизованный доступ"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка при выходе из системы"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/available-cars": {
            "get": {
                "tags": [
                    "Автомобили"
                ],
                "summary": "Список автомобилей, доступных текущему пользователю между двумя временными метками",
                "operationId": "c325f74407334052b788b227f67d7c60",
                "parameters": [
                    {
                        "name": "start_time",
                        "in": "query",
                        "description": "Время начала периода",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "end_time",
                        "in": "query",
                        "description": "Время окончания периода",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "car_model_id",
                        "in": "query",
                        "description": "ID модели автомобиля",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "comfort_category_id",
                        "in": "query",
                        "description": "ID категории комфорта",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список доступных автомобилей",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CarResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Аутентификация"
                ],
                "summary": "Получить данные текущего пользователя",
                "description": "Возвращает данные текущего авторизованного пользователя",
                "operationId": "8c6eb38d8f7bd492c2c8fc1ed239a6b2",
                "responses": {
                    "200": {
                        "description": "Успешное выполнение",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Неавторизованный доступ"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "CarResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "license_plate": {
                        "type": "string",
                        "example": "USA-001"
                    },
                    "vin": {
                        "type": "string",
                        "example": "1GYS4DKJ0MR123456",
                        "nullable": true
                    },
                    "manufactured_year": {
                        "type": "integer",
                        "example": 2023,
                        "nullable": true
                    },
                    "color": {
                        "type": "string",
                        "example": "Black",
                        "nullable": true
                    },
                    "model": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 10
                            },
                            "name": {
                                "type": "string",
                                "example": "Cadillac Escalade"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "comfort_category": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 99
                            },
                            "name": {
                                "type": "string",
                                "example": "Class A"
                            },
                            "level": {
                                "type": "integer",
                                "example": 1
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "driver": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 5
                            },
                            "full_name": {
                                "type": "string",
                                "example": "Michael Carter"
                            },
                            "phone": {
                                "type": "string",
                                "example": "+1 206 555 0101",
                                "nullable": true
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Olivia Bennett"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "executive@example.com"
                    },
                    "position": {
                        "type": "string",
                        "example": "Chief Executive Officer",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use JWT token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Аутентификация",
            "description": "Аутентификация"
        },
        {
            "name": "Автомобили",
            "description": "Автомобили"
        }
    ]
}